LOLLIPOP-MLCR-PATCHSET: Add patches for support of master perf
With the patches added here, we can run the perf command from
master branch on the Lollipop builds with the the lollipo
version bionic.

the patches added with this change are on projects of bionic,
external/elfutils, external/linux-tools-perf.
the elfuitls and linux-tools-perf projects are from master branch,
and the bionic is based on the Lollipop version.

Change-Id: I89166945e0eab09f40f1ca3c425347e8bc7d75db
Signed-off-by: Yongqin Liu <yongqin.liu@linaro.org>