hikey-optee-*-pullreqs: remove condition on..
ghprbPullId. Since we've disabled patching with pull-request in
ebf3441e, these patches now just serves to resync the optee_* repos
to master, but they're still conditioned on the existence of a
ghprbPullId. Remove this condition so that the patches can be applied
properly.

Changed some comments to actual echo statements along the way.

Change-Id: I691127c9390fcfeaf322bbdf7d55e338342495f5
Signed-off-by: Victor Chong <victor.chong@linaro.org>