lkft-mainline: remove the revert of "security: Implement Clang's stack initialization"
After the change of [1], we need to specify the CLang options when run make hikey_defconfig,
Otherwise, gcc will be used for .config generation, and cause difference
when run make Image, which would cause a prompt to be popped up, and need manual input.

With CLANG options specified for the make hikey_defconfig command, we don't need to
revert this chanage any more.

[1]: https://android.googlesource.com/kernel/common/+/709a972efb01efaeb97cad1adc87fe400119c8ab%5E%21/

Signed-off-by: Yongqin Liu <yongqin.liu@linaro.org>
Change-Id: I9cae8a7e2798b822d159923ab1d0ace6c497cba0