optee: chgrev: change function declaration format
Jenkins doesn't seem to like the current function declaration format so
follow those in 'functions' file. Perhaps the bash version in the
Jenkins builders is different than those tested locally.

Fix errors like in:
https://ci.linaro.org/job/android-hikey-optee-p/284
https://ci.linaro.org/job/android-hikey-optee-p/285

./android-patchsets/hikey-hikey960-optee-chgrev-common: 18: ./android-patchsets/hikey-hikey960-optee-chgrev-common: Syntax error: "(" unexpected

Note that there was also another error with using "else if" instead of
elif in 284 that wasn't 'detected' (i.e. not shown in the logs) due to
this error (function format) being detected first and stopping the job
before it could be detected. That error has been fixed in a previous
commit: 7cce05ff ("optee: chgrev: use elif not "else if" in bash")

Change-Id: I1ebc5a543b0e9022802e925727adb86de07ac6eb
Signed-off-by: Victor Chong <victor.chong@linaro.org>